# app/Models/DocField.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

/**
 * @OA\Schema()
 */
class DocField extends Model
{
    use HasFactory;

    protected $guarded = false;
}