# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doc-Configurator API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.3.0"
    },
    "servers": [
        {
            "url": "https://guehakosu.beget.app/api/v1/",
            "description": "Doc-Configurator API host"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register new User",
                "operationId": "userRegister",
                "requestBody": {
                    "description": "Login user",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "pass_confirm": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login to authenticate User",
                "operationId": "userLogin",
                "requestBody": {
                    "description": "Login user",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid email/password supplied"
                    }
                }
            }
        },
        "/doc_fields": {
            "get": {
                "tags": [
                    "DocField"
                ],
                "summary": "Read",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "DocField not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "DocField"
                ],
                "summary": "Create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "doc_field_name": {
                                                "type": "string",
                                                "example": "Some field name"
                                            },
                                            "doc_field_id": {
                                                "type": "integer",
                                                "example": 33
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "doc_field_name": {
                                            "type": "string",
                                            "example": "Some field name"
                                        },
                                        "doc_field_id": {
                                            "type": "integer",
                                            "example": 33
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "DocTemplate not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/doc_fields/{id}": {
            "put": {
                "tags": [
                    "DocField"
                ],
                "summary": "Update",
                "operationId": "updateDocField",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of DocField to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "DocField not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "DocField"
                ],
                "summary": "Delete",
                "operationId": "deleteDocField",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "DocField ID to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Deleted"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "DocField not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/doc_templates": {
            "get": {
                "tags": [
                    "DocTemplates"
                ],
                "summary": "Get all DocTemplates",
                "operationId": "getDocTemplates",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DocTemplate"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "DocTemplate not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "DocTemplates"
                ],
                "summary": "Get all DocTemplates",
                "operationId": "42c2795e1196710c0eebd5000e0d3ab3",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DocTemplate"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "DocTemplate not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/doc_templates/{id}": {
            "get": {
                "tags": [
                    "DocTemplates"
                ],
                "summary": "Get DocTemplate by ID",
                "description": "Returns a single DocTemplate",
                "operationId": "getDocTemplateById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of DocTemplate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "DocTemplate not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "DocTemplates"
                ],
                "summary": "Update an existing DocTemplate by ID",
                "operationId": "updateDocTemplate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of DocTemplate to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "DocTemplate not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "DocTemplates"
                ],
                "summary": "Deletes a DocTemplate",
                "operationId": "deleteDocTemplate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "DocTemplate ID to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Deleted"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "DocTemplate not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "DocField": {},
            "DocTemplate": {},
            "Document": {},
            "User": {}
        }
    }
}